requirements.txt
transformers==4.44.2
openai==1.42.0
pandas==2.1.4
python-dotenv==1.0.1